+++ src/routing/Link.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { forwardRef } from "react";
 
 import { AnyRoute, Route, RouteArguments } from "./createRoute";
 import { useHref } from "./useHref";
@@ -33,34 +33,70 @@ export type LinkProps<
    * onClick works as normal, but if you return false from it the navigation will not happen
    */
   onClick?: (e: React.MouseEvent<HTMLAnchorElement>) => boolean | void;
+  preloadOnInteraction?: boolean;
+  preloadOnHoverMs?: number;
 } & RouteArguments<TRouteParams, TRouteQuery, TRouteMeta> &
   Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, "href" | "onClick">;
 
-/**
- * @public
- *
- * Renders an anchor tag pointing at the provided Route
- *
- * The query/params/meta props are conditionally required based on the
- * route passed as the To parameter
- */
-export function Link<TRoute extends AnyRoute>({
-  to,
-  children,
-  testId,
-  ...rest
-}: LinkProps<TRoute>) {
+function LinkInner<TRoute extends AnyRoute>(
+  {
+    to,
+    children,
+    testId,
+    preloadOnHoverMs,
+    preloadOnInteraction,
+    onMouseDown: _onMouseDown,
+    onMouseEnter: _onMouseEnter,
+    onMouseLeave: _onMouseLeave,
+    ...rest
+  }: LinkProps<TRoute>,
+  ref: React.ForwardedRef<HTMLAnchorElement>
+) {
   // @ts-ignore, these fields _might_ exist, so typechecking doesn't believe they exist
   // and everything that consumes params/query already checks for undefined
   const { params, query, meta, ...props } = rest;
 
+  let timeout: ReturnType<typeof setTimeout> | undefined;
   const href = useHref(to, params, query);
+  const onMouseDown: React.MouseEventHandler<HTMLAnchorElement> | undefined =
+    preloadOnInteraction
+      ? (e) => {
+          _onMouseDown?.(e);
+
+          to.preload({ params, query, meta });
+        }
+      : undefined;
+  const onMouseEnter: React.MouseEventHandler<HTMLAnchorElement> | undefined =
+    preloadOnHoverMs !== undefined
+      ? (e) => {
+          _onMouseEnter?.(e);
+
+          timeout = setTimeout(() => {
+            to.preload({ params, query, meta });
+          }, preloadOnHoverMs);
+        }
+      : undefined;
+  const onMouseLeave: React.MouseEventHandler<HTMLAnchorElement> | undefined =
+    preloadOnHoverMs !== undefined
+      ? (e) => {
+          _onMouseLeave?.(e);
+
+          if (timeout !== undefined) {
+            clearTimeout(timeout);
+          }
+        }
+      : undefined;
 
   return (
     <a
       {...props}
+      ref={ref}
       href={href}
       data-testid={testId}
+      onMouseDown={onMouseDown ?? _onMouseDown}
+      onMouseEnter={onMouseEnter ?? _onMouseEnter}
+      onMouseLeave={onMouseLeave ?? _onMouseLeave}
+      data-xstate-tree
       onClick={(e) => {
         if (props.onClick?.(e) === false) {
           return;
@@ -73,7 +109,6 @@ export function Link<TRoute extends AnyRoute>({
         }
 
         e.preventDefault();
-
         to.navigate({ params, query, meta });
       }}
     >
@@ -81,3 +116,15 @@ export function Link<TRoute extends AnyRoute>({
     </a>
   );
 }
+
+/**
+ * @public
+ *
+ * Renders an anchor tag pointing at the provided Route
+ *
+ * The query/params/meta props are conditionally required based on the
+ * route passed as the To parameter
+ */
+export const Link = forwardRef(LinkInner) as <TRoute extends AnyRoute>(
+  props: LinkProps<TRoute> & { ref?: React.ForwardedRef<HTMLAnchorElement> }
+) => ReturnType<typeof LinkInner>;